Olistay-RPCServer/requirements.txt
# OLISTAY Recommendation Server Dependencies

# Core RPC Framework
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# Web Framework (Alternative to gRPC)
flask==3.0.0
flask-cors==4.0.0

# Data Processing & ML

pandas==2.1.4
scikit-learn==1.3.2

# Database Drivers
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1

# Caching & Performance
hiredis==2.3.2

# Elasticsearch
elasticsearch==8.11.1

# Environment Management
python-dotenv==1.0.0

# Logging & Monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Production Server
gunicorn==21.2.0
uvicorn==0.25.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2